requirements.txt
Pillow==9.5.0
PyYAML==6.0 